feat:add emoji utils and move emoji functions
